requirements.txt
streamlit
python-docx
pdfplumber
scikit-learn